Terminating this branch. And we move to 2D.
According to the book there are a lot of stuff to do in 1D. Like Foirier
analysis, of some vivlet and incorporation of 2nd order ABC based on
differential operator of absorbation derived from solving wave quation
at the Border.
But lets just skip to 2D and try to design out programm.